- Improve behavior uplink source combo. Removed source will automatically disposed from items. - Fix crash on switching scene collection during output is active.
